+++ lib/core/viewmodels/main_page/chat/chat_inside_view_model.dart
@@ -149,6 +149,7 @@ class ChatInsideViewModel extends BaseViewModel {
       }
       lastDatePartition?.add([message]);
     }
+
     return partitions;
   }
