+++ lib/core/viewmodels/main_page/feed/reaction_view_model.dart
@@ -2,35 +2,28 @@
 // Copyright 2025 BitCodersNN
 
 import 'package:injector/injector.dart';
-import 'package:unn_mobile/core/misc/user/current_user_sync_storage.dart';
 import 'package:unn_mobile/core/models/feed/rating_list.dart';
 import 'package:unn_mobile/core/models/profile/user_short_info.dart';
 import 'package:unn_mobile/core/services/interfaces/feed/legacy/getting_rating_list.dart';
 import 'package:unn_mobile/core/services/interfaces/feed/legacy/getting_vote_key_signed.dart';
 import 'package:unn_mobile/core/services/interfaces/feed/reaction_service.dart';
-import 'package:unn_mobile/core/viewmodels/base_view_model.dart';
 import 'package:unn_mobile/core/viewmodels/factories/reaction_view_model_factory.dart';
+import 'package:unn_mobile/core/viewmodels/main_page/common/reaction_view_model_base.dart';
 
-class ReactionViewModel extends BaseViewModel {
-  static const noId = -1;
-
+class ReactionViewModel extends ReactionViewModelBase {
   final GettingVoteKeySigned _gettingVoteKeySigned;
   final GettingRatingList _gettingRatingList;
   final ReactionService _reactionManager;
-  final CurrentUserSyncStorage _currentUserSyncStorage;
 
-  RatingList? _ratingList;
   String? _voteKeySigned;
 
   bool _isLoading = true;
-
-  bool _isChangingReaction = false;
 
   ReactionViewModel(
     this._gettingVoteKeySigned,
     this._gettingRatingList,
     this._reactionManager,
-    this._currentUserSyncStorage,
+    super._currentUserSyncStorage,
   );
 
   factory ReactionViewModel.cached(ReactionCacheKey key) {
@@ -39,34 +32,7 @@ class ReactionViewModel extends BaseViewModel {
         .getViewModel(key);
   }
 
-  bool get canAddReaction =>
-      ReactionType.values
-          .map((e) => getReactionCount(e))
-          .where((e) => e > 0)
-          .length <
-      ReactionType.values.length;
-
-  ReactionType? get currentReaction {
-    return _ratingList?.getReactionByUser(
-      _currentUserSyncStorage.currentUserData?.bitrixId ?? noId,
-    );
-  }
-
-  bool get isChangingReaction => _isChangingReaction;
-
   bool get isLoading => _isLoading;
-
-  int get reactionCount => _ratingList?.getTotalNumberOfReactions() ?? 0;
-
-  Iterable<ReactionType> get reactionList =>
-      _ratingList?.ratingList.entries
-          .where((entry) => entry.value.isNotEmpty)
-          .map((e) => e.key) ??
-      [];
-
-  int getReactionCount(ReactionType reaction) {
-    return _ratingList?.getNumberOfReactions(reaction) ?? 0;
-  }
 
   void init({String? voteKeySigned, int? postId, int? authorId}) {
     _isLoading = true;
@@ -79,32 +45,10 @@ class ReactionViewModel extends BaseViewModel {
   }
 
   void initFull(String voteKeySigned, RatingList ratingList) {
-    _ratingList = ratingList;
+    this.ratingList = ratingList;
     _voteKeySigned = voteKeySigned;
     _isLoading = false;
     notifyListeners();
-  }
-
-  void toggleLike() {
-    if (currentReaction != null) {
-      _setReaction(null);
-    } else {
-      _setReaction(ReactionType.like);
-    }
-    notifyListeners();
-  }
-
-  void toggleReaction(ReactionType reaction) async {
-    if (_isChangingReaction) {
-      return;
-    }
-    final currentReaction = this.currentReaction;
-
-    // Убираем реакцию и ставим новую, если она не та же, что была
-    await _setReaction(null);
-    if (currentReaction != reaction) {
-      await _setReaction(reaction);
-    }
   }
 
   Future<void> _loadData({
@@ -122,66 +66,56 @@ class ReactionViewModel extends BaseViewModel {
     if (_voteKeySigned == null) {
       return;
     }
-    _ratingList = await _gettingRatingList.getRatingList(
+    ratingList = await _gettingRatingList.getRatingList(
       voteKeySigned: _voteKeySigned!,
     );
   }
 
-  Future<void> _setReaction(ReactionType? reaction) async {
-    if (_isChangingReaction) {
+  @override
+  Future<void> setReaction(ReactionType? reaction) async {
+    if (profileId == ReactionViewModelBase.noId) {
       return;
     }
-    try {
-      _isChangingReaction = true;
-      final profileId = _currentUserSyncStorage.currentUserData?.bitrixId;
-      if (profileId == null) {
+    if (_voteKeySigned == null) {
+      return;
+    }
+    if (reaction == null) {
+      final currentReactionInfo = ratingList?.getReactionInfoByUser(profileId);
+      final currentReactionType = ratingList?.getReactionByUser(profileId);
+      if (currentReactionInfo == null || currentReactionType == null) {
         return;
       }
-      if (_voteKeySigned == null) {
-        return;
-      }
-      if (reaction == null) {
-        final currentReactionInfo =
-            _ratingList?.getReactionInfoByUser(profileId);
-        final currentReactionType = _ratingList?.getReactionByUser(profileId);
-        if (currentReactionInfo == null || currentReactionType == null) {
-          return;
-        }
-        // Временно удаляем реакцию, чтобы показать действие
-        _ratingList!.removeReaction(profileId);
-        notifyListeners();
-        if (!await _reactionManager.removeReaction(_voteKeySigned!)) {
-          // Если реакция не удалилась - восстанавливаем её
-          _ratingList!.addReactions(
-            currentReactionType,
-            [currentReactionInfo],
-          );
-          notifyListeners();
-        }
-      } else {
-        // Добавляем временно, чтобы сразу показать действие
-        _ratingList?.addReactions(
-          reaction,
-          [UserShortInfo(profileId, '', '')],
+      // Временно удаляем реакцию, чтобы показать действие
+      ratingList!.removeReaction(profileId);
+      notifyListeners();
+      if (!await _reactionManager.removeReaction(_voteKeySigned!)) {
+        // Если реакция не удалилась - восстанавливаем её
+        ratingList!.addReactions(
+          currentReactionType,
+          [currentReactionInfo],
         );
         notifyListeners();
-        final reactionUserInfo = await _reactionManager.addReaction(
-          reaction,
-          _voteKeySigned!,
-        );
-        // Удаляем временную реакцию
-        _ratingList?.removeReaction(profileId);
-        if (reactionUserInfo != null) {
-          // Если реакция реально добавилась - фиксируем это
-          _ratingList?.addReactions(
-            reaction,
-            [reactionUserInfo],
-          );
-        }
-        notifyListeners();
       }
-    } finally {
-      _isChangingReaction = false;
+    } else {
+      // Добавляем временно, чтобы сразу показать действие
+      ratingList?.addReactions(
+        reaction,
+        [UserShortInfo(profileId, '', '')],
+      );
+      notifyListeners();
+      final reactionUserInfo = await _reactionManager.addReaction(
+        reaction,
+        _voteKeySigned!,
+      );
+      // Удаляем временную реакцию
+      ratingList?.removeReaction(profileId);
+      if (reactionUserInfo != null) {
+        // Если реакция реально добавилась - фиксируем это
+        ratingList?.addReactions(
+          reaction,
+          [reactionUserInfo],
+        );
+      }
       notifyListeners();
     }
   }
