+++ lib/load_services.dart
@@ -132,6 +132,7 @@ import 'package:unn_mobile/core/services/interfaces/common/search_id_on_portal_s
 import 'package:unn_mobile/core/services/interfaces/common/storage_service.dart';
 import 'package:unn_mobile/core/providers/interfaces/profile/user_data_provider.dart';
 import 'package:unn_mobile/core/viewmodels/auth_page/auth_page_view_model.dart';
+import 'package:unn_mobile/core/viewmodels/factories/message_reaction_view_model_factory.dart';
 import 'package:unn_mobile/core/viewmodels/main_page/certificates_online/certificate_item_view_model.dart';
 import 'package:unn_mobile/core/viewmodels/main_page/certificates_online/certificates_view_model.dart';
 import 'package:unn_mobile/core/viewmodels/factories/attached_file_view_model_factory.dart';
@@ -616,6 +617,9 @@ void registerDependencies() {
 
   injector.registerSingleton<ReactionViewModelFactory>(
     () => ReactionViewModelFactory(),
+  );
+  injector.registerSingleton<MessageReactionViewModelFactory>(
+    () => MessageReactionViewModelFactory(),
   );
 
   injector.registerSingleton<FeedPostViewModelFactory>(
